+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Fixed
+
+- encode the object key of `x-tos-copy-source` header when the object'size is 0
+
 ## [2.1.16] - 2022-11-09
 
 ### Fixed