+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.20] - 2022-12-08
+
+### Added
+
+- add `downloadFile` method
+
 ## [2.1.19] - 2022-11-25
 
 ### Added